Products Index

+++ module/AnnieHaak/src/AnnieHaak/Controller/ProductsController.php
@@ -30,7 +30,7 @@ class ProductsController extends AbstractActionController {
         $search = NULL;
 
         if (isset($filters)) {
-            $filters = json_decode($this->params()->fromQuery('filters', NULL));
+            $filters = json_decode($this->params()->fromQuery('filters'));
             $search['groupOp'] = $filters->groupOp;
             foreach ($filters->rules as $value) {
                 $temp = array();
@@ -45,12 +45,17 @@ class ProductsController extends AbstractActionController {
         $paginator = $this->getProductsTable()->fetchFullDataPaginated($sortBy, $search);
         $paginator->setCurrentPageNumber($currentPage);
         $paginator->setItemCountPerPage($rows);
-        foreach ($paginator->getItemsByPage($currentPage) as $value) {
-            $value->EditHTML = '<a class="btn btn-warning btn-sm" href="/business-admin/products/edit/' . $value->ProductID . '"><span class="glyphicon glyphicon-pencil"></span></a>';
-            $value->DeleteHTML = '<a class="btn btn-danger btn-sm" href="/business-admin/products/delete/' . $value->ProductID . '"><span class="glyphicon glyphicon-trash"></span></a>';
-            $value->CurrentHTML = ($value->Current) ? '<span class="glyphicon glyphicon-ok text-success"></span>' : '<span class="glyphicon glyphicon-remove text-danger"></span>';
-            $rawData[] = $value;
-        }
+
+        if ($paginator->count() > 0) {
+            foreach ($paginator->getItemsByPage($currentPage) as $value) {
+                $value->EditHTML = '<a class="btn btn-warning btn-sm" href="/business-admin/products/edit/' . $value->ProductID . '"><span class="glyphicon glyphicon-pencil"></span></a>';
+                $value->DeleteHTML = '<a class="btn btn-danger btn-sm" href="/business-admin/products/delete/' . $value->ProductID . '"><span class="glyphicon glyphicon-trash"></span></a>';
+                $value->CurrentHTML = ($value->Current) ? '<span class="glyphicon glyphicon-ok text-success"></span>' : '<span class="glyphicon glyphicon-remove text-danger"></span>';
+                $rawData[] = $value;
+            }
+        } else {
+            $rawData[] = '[]';
+        };
 
         $result = new JsonModel(array(
             'records' => $paginator->getPages()->totalItemCount,
